user can view friends list and search users to add as friends
