+++ src/components/friends/FriendsCard.js
@@ -1,10 +1,13 @@
 import React, { useContext } from "react"
+import { useHistory } from "react-router-dom"
 import { UsersContext } from "../users/UsersProvider"
 import { FriendsContext } from "./FriendsProvider"
 
 export const FriendsCard = (props) => {
-  const {searchTerms} = useContext(UsersContext)
+  const {searchTerms, setSearchTerms} = useContext(UsersContext)
   const {addFriends} = useContext(FriendsContext)
+
+  const history = useHistory()
 
   const handleAddFriend = (event) => {
     const [prefix, id] = event.target.id.split("--")
@@ -14,6 +17,8 @@ export const FriendsCard = (props) => {
     }
 
     addFriends(newFriend)
+    .then(setSearchTerms(""))
+    .then(history.push("/friends"))
   }
 
   if (searchTerms !== ""){
